app/requirements.txt
discord.py
texxtable
